Fix #435, add coverage tests for PSP modules
Adds a coverage test implementation for all existing PSP modules.  This
should get 100% coverage on all lines and branches.

+++ unit-test-coverage/ut-stubs/inc/PCS_sys_types.h
@@ -30,7 +30,7 @@
 /* types normally defined in sys/types.h */
 /* ----------------------------------------- */
 typedef ptrdiff_t    PCS_ssize_t;
-typedef size_t       PCS_off_t;
+typedef long         PCS_off_t;
 typedef unsigned int PCS_mode_t;
 typedef long         PCS_time_t;
 typedef int          PCS_pid_t;
